chore: sync uncommitted Rust source changes (nightshift 2026-03-17)

+++ crates/ruvector-solver/src/neumann.rs
@@ -169,7 +169,7 @@ impl NeumannSolver {
                 break;
             }
             for j in 0..n {
-                v[j] = w[j] / w_norm as f32;
+                v[j] = w[j] / w_norm;
             }
         }
